Refactor Owner form in Edit and Create Owner
Related to issue #4
